Implementado proceso de compra
* Implementado seleccionar direccion o crearla
* Implementado seleccionar tarjeta de crédito o crearla
* Implementadas vistas proceso de compra
* Creados ficheros en config para almacenar provincias y paises
* Creadas rutas necesarias para todo el proceso
* Añadidos a modelos campos fillable necesarios

+++ my-project/routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AddressesController;
 use App\Http\Controllers\CartController;
 use App\Http\Controllers\ProductsController;
 use Illuminate\Support\Facades\Route;
@@ -23,9 +24,15 @@ Route::get('/', function () {
 Route::get('/productos/{id}', [ProductsController::class, 'show'])->name('products.show');
 
 // Cart
-Route::get('/cesta', [CartController::class, 'index'])->name('cart.index');
-Route::get('/cesta/decrementar/{id}', [CartController::class, 'decreaseProduct'])->name('cart.decrease');
-Route::get('/cesta/incrementar/{id}', [CartController::class, 'increaseProduct'])->name('cart.increase');
-Route::post('/cesta', [CartController::class, 'createItemCart'])->name('cart.add');
-Route::delete('/cesta/eliminar/{id}', [CartController::class, 'deleteItemCart'])->name('cart.delete');
-Route::post('/cesta/discount', [CartController::class, 'applyDiscount'])->name('cart.apply_discount');
+Route::get('/cesta', [CartController::class, 'index'])->name('cart.index')->middleware('auth');
+Route::post('/cesta/comprar', [CartController::class, 'buyProccess'])->name('cart.buy_proccess')->middleware('auth');
+Route::get('/cesta/direccion', [CartController::class, 'shipping'])->name('cart.shipping')->middleware('auth');
+Route::post('/cesta/direccion/seleccionar', [CartController::class, 'shippingSelect'])->name('cart.shipping_select')->middleware('auth');
+Route::get('/cesta/pago', [CartController::class, 'payMethod'])->name('cart.pay_method')->middleware('auth');
+Route::post('/cesta/pago/finalizar', [CartController::class, 'pay'])->name('cart.pay')->middleware('auth');
+Route::get('/cesta/confirmacion', [CartController::class, 'finish'])->name('cart.finish')->middleware('auth');
+Route::get('/cesta/decrementar/{id}', [CartController::class, 'decreaseProduct'])->name('cart.decrease')->middleware('auth');
+Route::get('/cesta/incrementar/{id}', [CartController::class, 'increaseProduct'])->name('cart.increase')->middleware('auth');
+Route::post('/cesta', [CartController::class, 'createItemCart'])->name('cart.add')->middleware('auth');
+Route::delete('/cesta/eliminar/{id}', [CartController::class, 'deleteItemCart'])->name('cart.delete')->middleware('auth');
+Route::post('/cesta/discount', [CartController::class, 'applyDiscount'])->name('cart.apply_discount')->middleware('auth');
